WIP some improvements
