+++ ui/src/app/utils/access-utils.tsx
@@ -259,9 +259,6 @@ export const GetStartedButton = ({style = {marginLeft: '0.5rem'}}) => <Button
     location.replace('/');
   }}>Get Started</Button>;
 
-export const isExpiring = (expiration: number): boolean =>
-  getWholeDaysFromNow(expiration) <= serverConfigStore.get().config.accessRenewalLookback;
-
 const withInvalidDateHandling = date => {
   if (!date) {
     return 'Unavailable';
